more docker stuff

+++ simscript_thomas.py
@@ -88,8 +88,8 @@ def snrbatch(fch1,bwchan,nchan,tsamp,mode,label,nsamp,npulse,sigmarange,dmrange,
                 model.inject(base1/model.write_snr()[1]*ampl)
                 w.write(model.write_snr()[0][:-2]+";"+str(dynspec.L2_clean(base2/model.write_snr()[1]*50))+f";{xset}"+"\n")
                 # print(f" {np.round(i,1)}/11.1 completed", end = "\r")
-                print(round((j-dm_start)/step),round((i-sig_start)/sig_step))
-                SN_array[round((j-dm_start)/step),round((i-sig_start)/sig_step)]=float(dynspec.L2_clean(base2/model.write_snr()[1]*50))
+                print(int(round((j-dm_start)/step)),int(round((i-sig_start)/sig_step)))
+                SN_array[int(round((j-dm_start)/step)),int(round((i-sig_start)/sig_step))]=float(dynspec.L2_clean(base2/model.write_snr()[1]*50))
                 model.writenoise(nsamp=nsamp)
             model.writenoise(nsamp=nsamp) ## write noise
             model.closefile()
